FTP Upload and Edit Files

+++ ftp_file_to_domains.php
@@ -5,6 +5,7 @@
 ?>
 <!DOCTYPE html>
 <html lang="es">
+
 <head>
 	<meta charset="utf-8">
 	<meta http-equiv="X-UA-Compatible" content="IE=edge">
@@ -14,7 +15,7 @@
 	<!-- Normalize -->
 	<link href="css/normalize.css" rel="stylesheet">
 	<!-- Bootstrap -->
-	<link href="css/bootstrap.min.css" rel="stylesheet">
+	<link href="css/bootstrap.css" rel="stylesheet">
 	<!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->
 	<!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
 	<!--[if lt IE 9]>
@@ -22,16 +23,23 @@
 	  <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
 	<![endif]-->
 </head>
+
 <body>
 	<div id="container">
 		<?php
 		date_default_timezone_set("America/Mexico_City");
 		$date = date('Y/m/d H:i:s');
+		error_reporting(E_ALL);
+		ini_set('display_errors', 1);
+
 
 		echo "<p>Inició: $date</p>";
 
 		set_time_limit(0);
 		require('comun.php');
+		require('ftp_accounts.php');
+
+
 
 		// Files to replicate.
 		$files = array(
@@ -46,7 +54,7 @@
 			// "../pm/ci/application/views/componentes/marcas-carrusel.php" => "/ci/application/views/componentes/marcas-carrusel.php",
 			// "../pm/ci/application/views/pages/aside_marcas_slider.php" => "/ci/application/views/pages/aside_marcas_slider.php",
 			// "../pm/ci/application/views/pages/aside_requisitos.php" => "/ci/application/views/pages/aside_requisitos.php",
-			"../pm/ci/application/views/pages/aviso_de_privacidad.php" => "/ci/application/views/pages/aviso_de_privacidad.php",
+			//"../pm/ci/application/views/pages/aviso_de_privacidad.php" => "/ci/application/views/pages/aviso_de_privacidad.php",
 			// "../pm/ci/application/views/pages/chat.php" => "/ci/application/views/pages/chat.php",
 			// "../pm/ci/application/views/pages/contenedor_lateral.php" => "/ci/application/views/pages/contenedor_lateral.php",
 			// "../pm/ci/application/views/pages/contenedor_lateral_menu_categorias.php" => "/ci/application/views/pages/contenedor_lateral_menu_categorias.php",
@@ -71,9 +79,12 @@
 			// "../pm/ci/application/views/pages/sucursales.php" => "/ci/application/views/pages/sucursales.php",
 			// "../pm/ci/application/views/plantillas/p_categoria.php" => "/ci/application/views/plantillas/p_categoria.php",
 			// "../pm/ci/application/views/plantillas/p_equipo.php" => "/ci/application/views/plantillas/p_equipo.php",
+			"../pm/ci/application/views/errors/html/error_404.php" => "/ci/application/views/errors/html/error_404.php",
 
 			//  "../pm/ci/application/controllers/Maquinaria.php" => "/ci/application/controllers/Maquinaria.php",
 			// "../pm/ci/application/controllers/Parametros.php" => "/ci/application/controllers/Parametros.php",
+			"../pm/ci/application/controllers/Errors.php" => "/ci/application/controllers/Errors.php",
+
 
 			// "../pm/public/js/app.js" => "/public_html/js/app.js",
 			//  "../pm/public/css/bundle.css" => "/public_html/css/bundle.css",
@@ -96,57 +107,80 @@
 			// "../pm/public/imagenes/renta_maquinaria_02.jpg" => "/public_html/imagenes/renta_maquinaria_02.jpg",
 		);
 
-		foreach($domains as $domain){
-			$ftp_server = $domain[0];
-			$ftp_user_name = $domain[1];
-			$ftp_user_pass = $domain[2];
+		foreach ($domains as $domain) {
+			$host = $domain[0];
+			$user = $domain[1];
+			$pass = $domain[2];
+			$port = 21;
+			$timeout = 30;
+			$message = '';
 
-			// Connect to FTP.
-			$conn_id = ftp_connect($ftp_server);
-
-			// Login with username and password.
-			$login_result = ftp_login($conn_id, $ftp_user_name, $ftp_user_pass);
-
-			// turn passive mode on
-			ftp_pasv($conn_id, true);
-
-			// Check connection.
-			if ((!$conn_id) || (!$login_result)) {
-				echo "<h3><a target='_blank' href='http://www.$ftp_server'>$ftp_server</a> <span class='label label-danger'>Failed! <span class='glyphicon glyphicon-remove' aria-hidden='true'></span></span></h3>";
-				echo "<br /><br />";
-				continue;
-			} else {
-				echo "<h3><a  target='_blank' href='http://www.$ftp_server'>$ftp_server</a> <span class='label label-success'>Connected! <span class='glyphicon glyphicon-ok' aria-hidden='true'></span></span></h3>";
+			if (!function_exists('ftp_ssl_connect')) {
+				die("ftp_ssl_connect() not available. Ensure PHP FTP+OpenSSL is enabled.");
 			}
 
+			// Connect to FTP.
+			$conn = @ftp_ssl_connect($host, $port, $timeout);
+			if (!$conn) {
+				$message = "Could not connect securely to $host:$port";
+				echo "<h3><a target='_blank' href='http://www.$host'>$message</a> <span class='label label.danger'>Failed!</h3>";
+				echo "<br /><br />";
+				continue;
+			}
+
+			// Login with username and password.
+			$login = @ftp_login($conn, $user, $pass);
+			if (!$login) {
+				ftp_close($conn);
+				$message = "Login failed for $host";
+				echo "<h3><a target='_blank' href='http://www.$host'>$message</a> <span class='label label.danger'>Failed!</h3>";
+				echo "<br /><br />";
+				continue;
+			}
+			// After login, add these critical commands:
+			ftp_raw($conn, "PBSZ 0"); // Set protection buffer size
+			ftp_raw($conn, "PROT P"); // Encrypt data channel
+			ftp_pasv($conn, true);     // Enable passive mode
+			ftp_set_option($conn, FTP_USEPASVADDRESS, false);
+
+			echo "<h3><a  target='_blank' href='http://www.$host'>$host</a> <span class='label label-success'>Connected!</h3>";
 
 			// If connection ok, upload files.
 			echo "<table class='table table-striped table-hover'>\n";
 			echo "<thead><tr><th>Source File</th><th>Destination File</th><th>Copy Status</th></tr></thead>\n";
-			foreach($files as $source_file => $destination_file){
+			foreach ($files as $source_file => $destination_file) {
+
+				if (!file_exists($source_file)) {
+					echo "<tr class='danger'>";
+					echo "<td>$source_file</td><td></td><td>No existe archivo local</td></tr>";
+					continue;
+				}
+
+				// If remote file exists: backup file
+				// Check if the file exists
+				$file_size = ftp_size($conn, $destination_file);
+				if ($file_size != -1) {
+					$backup_name = $destination_file . "__" . date('Y-m-d_h:ia') . "replaced.bk";
+					$rename = ftp_rename($conn, $destination_file, $backup_name);
+				}
+
 				// Upload the file.
-
-				$f_exists = file_exists( $source_file);
-
-				$backup_name = $destination_file . "." . date('Y-m-d_h:ia') . ".bk";
-				$rename = ftp_rename($conn_id, $destination_file, $backup_name );
-				$upload = ftp_put($conn_id, $destination_file, $source_file, FTP_BINARY);
+				$upload = ftp_put($conn, $destination_file, $source_file, FTP_BINARY);
 
 				// Check upload status.
 				if (!$upload) {
 					echo "<tr class='danger'>";
-					echo "<td>$source_file</td><td>$destination_file</td><td><span class='glyphicon glyphicon-remove' aria-hidden='true'></span> FTP upload has failed!</td>";
+					echo "<td>$source_file</td><td>$destination_file</td><td> FTP upload has failed!</td></tr>";
 
-				}
-				else {
+				} else {
 					echo "<tr class='success'>";
-					echo "<td>$source_file</td><td>$destination_file</td><td><span class='glyphicon glyphicon-ok' aria-hidden='true'></span> Uploaded successfully</td>";
+					echo "<td>$source_file</td><td>$destination_file</td><td> Uploaded successfully</td></tr>";
 				}
-				echo "</tr>";
 			}
 			echo "</table>";
+
 			// Close the FTP stream.
-			ftp_close($conn_id);
+			ftp_close($conn);
 			echo "<br /><br />";
 
 		}
@@ -163,6 +197,5 @@
 	<!-- Include all compiled plugins (below) ================================== -->
 	<script src="js/bootstrap/bootstrap.min.js"></script>
 </body>
+
 </html>
-
-
